refactor: refactoring the database package
